% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3learners.randomforestsrc-package}
\alias{mlr3learners.randomforestsrc}
\alias{mlr3learners.randomforestsrc-package}
\title{mlr3learners.randomforestsrc: Learners from the {randomForestSRC} package for 'mlr3'}
\description{
Adds learners from the {randomForestSRC} package to
    'mlr3'.
}
\author{
\strong{Maintainer}: Raphael Sonabend \email{raphael.sonabend.15@ucl.ac.uk}

}
